add useValidationRuleRegistration hook, add RangeRule
