+++ packages/react/src/components/validation/index.ts
@@ -1,5 +1,5 @@
 export { ValidationGroup } from './components/validation-group';
-export { CustomRule } from './components/validation-rules';
+export { CustomRule, RangeRule } from './components/validation-rules';
 export { Validator } from './components/validator';
 export { ValidationEngineContext } from './contexts/validation-engine-context';
 export { ValidationGroupContext } from './contexts/validation-group-context';
